+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/Pages/Components/_Imports.razor
@@ -1,0 +1,1 @@
+﻿@inherits PerformanceLoggingComponentBase
