+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/Shared/Components/ComponentApiDoc.razor.cs
@@ -20,24 +20,11 @@ public partial class ComponentApiDoc
 	[Parameter] public bool Delegate { get; set; }
 
 	[Inject] protected IComponentApiDocModelBuilder ComponentApiDocModelBuilder { get; set; }
-	[Inject] protected ILogger<ComponentApiDoc> Logger { get; set; }
 
 	private ComponentApiDocModel model;
-	private TimeSpan? startTime;
 
 	protected override void OnParametersSet()
 	{
-		startTime ??= DateTime.Now.TimeOfDay;
-
-		model = ComponentApiDocModelBuilder.BuildModel(this.Type, this.Delegate);
-
-		Logger.LogWarning($"ComponentApiDoc({this.GetHashCode()})_DownloadFileAndGetSummaries: Elapsed {(DateTime.Now.TimeOfDay - startTime.Value).TotalMilliseconds} ms");
-	}
-
-
-	protected override void OnAfterRender(bool firstRender)
-	{
-		Logger.LogWarning($"ComponentApiDoc({this.GetHashCode()})_OnAfterRender: Elapsed {(DateTime.Now.TimeOfDay - startTime.Value).TotalMilliseconds} ms");
-		startTime = null;
+		model ??= ComponentApiDocModelBuilder.BuildModel(this.Type, this.Delegate);
 	}
 }
