+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/Shared/Components/Demo.razor
@@ -1,0 +1,46 @@
+﻿@using Havit.Blazor.Components.Web.Bootstrap.Documentation.Pages
+@inherits PerformanceLoggingComponentBase
+
+<!--googleoff: index-->
+
+<div class="card card-demo my-3">
+
+    @if (Tabs)
+    {
+        <div class="card-header">
+            <HxTabPanel CssClass="card-header-tabs">
+                <HxTab Id="demo" Title="Demo" OnTabActivated="HandleDemoTabActivated" />
+                <HxTab Id="source" Title="Source" OnTabActivated="HandleCodeTabActivated" />
+            </HxTabPanel>
+        </div>
+    }
+
+    <div class="card-body p-0">
+
+        @if (showingDemo || !Tabs)
+        {
+            <div class="@CssClassHelper.Combine("p-3", this.DemoCardCssClass)">
+                <DynamicComponent Type="this.Type" />
+            </div>
+        }
+        @if (!showingDemo || !Tabs)
+        {
+            if (code is null)
+            {
+                <HxPlaceholderContainer Animation="PlaceholderAnimation.Glow">
+                    <HxPlaceholder Columns="8" />
+                </HxPlaceholderContainer>
+            }
+            else
+            {
+                <pre class="@(!Tabs ? "gray-background" : null)">
+                    <code class="language-cshtml">
+                        @(code.Trim())
+                    </code>
+                </pre>
+            }
+        }
+    </div>
+</div>
+
+<!--googleon: index-->
