+++ Havit.Blazor.Components.Web.Bootstrap.Documentation/XmlDoc/Havit.Blazor.Components.Web.xml
@@ -605,7 +605,7 @@
         <member name="T:Havit.Blazor.Components.Web.HxRepeater`1">
             <summary>
             A data-bound list component that allows custom layout by repeating a specified template for each item displayed in the list.
-            Analogous to ASP.NET WebForms Repater control.
+            Analogous to ASP.NET WebForms Repeater control.
             </summary>
             <typeparam name="TItem">item type</typeparam>
         </member>
