+++ Havit.Blazor.Components.Web/HxRepeater.razor.cs
@@ -2,7 +2,7 @@
 {
 	/// <summary>
 	/// A data-bound list component that allows custom layout by repeating a specified template for each item displayed in the list.
-	/// Analogous to ASP.NET WebForms Repater control.
+	/// Analogous to ASP.NET WebForms Repeater control.
 	/// </summary>
 	/// <typeparam name="TItem">item type</typeparam>
 	public partial class HxRepeater<TItem> : ComponentBase
